tg-update.sh: avoid false positive git checkout complaints
When attempting to do a `git checkout` and there are untracked
files in the current worktree which would be overwritten by
the commit about to be checked out, Git complains and aborts
the checkout unless extra options are given that tell it to
go ahead and step on the untracked files in the worktree.

Unfortunately, Git is too stupid to check first to see if
what is about to be checked out exactly matches what's already
in the worktree as an untracked file.  Usually Git is better
about things like this.

Attempt to compensate for Git's poor behavior and detect
exactly matching untracked files that are about to be stepped
on with exactly the same contents (including the file mode)
and allow the checkout to proceed in that case without complaint.

Provide this support for regular files and symbolic links as
long as they do not contain any '\n' characters in their names
(the commands used to perform the tests would get confused in
that case -- spaces and tabs are handled just fine).

This situation seems somehow much more likely to occur when
working with various TopGit branches some of which contain a
specific file and some of which do not and at the same time
creating new TopGit branches based on combinations of those
other branches.

For the time being, this support is limited to the post-update
checkout where TopGit returns to the original starting point
before the `tg update` command was given.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>