t/t*.sh: do not ignore failures in for loops
With a test that does something like:

    for x in a b c d; do
      test_for_failure $x
    done

Expects that a failure testing a, b, or c will be detected, it
loses.  Only the result of the last command executed within the
loop will make it out and that would be, in this case, the test
of d.  Instead, change the loop like so:

    for x in a b c d; do
      test_for_failure $x || return
    done

The "return" statement uses the value of $? if a value is not
provided and the testing framework explicitly allows returning from
a unit test via return.  Errors thus result in a prompt exit with
a non-zero status code.

Make this change for affected tests to prevent them from missing
any unit test failures.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>