tg-push: make all pushes atomic
Replace the use of xargs with a temporary file that allows all the
branches being pushed to be passed to git at once regardless of how
many of them there are or how long their branch names are.

When giving the --signed and/or --atomic options this change makes
sure that only one signature will be required and that the push
truly is atomic.

Update the docs to remove the warning since it no longer applies.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>