t1060: update rev-parse --git-common-dir test
Although `git rev-parse --git-common-dir` was, indeed, broken,
it's since been fixed.

The test that was checking for brokenness, never having experienced
an unbroken version was unprepared for it to start working.

Update the test to pass when running a version of Git with a
`git rev-parse --git-common-dir` that's not actually broken.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>