tg-revert.sh: make -h output accessible without tg.sh
The output of -h needs to be available by running the .sh script
directly as it's used to create various help files.

It must not depend on having variables from tg.sh set.

Remodel the tg-revert.sh help to use the same technique tg-info.sh
does.  It's a bit (still needs work) more readable and no longer
uses any variables that may not be available.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>