tg-update: keep state in tg-state directory
Forthcoming updates will allow other commands to be interrupted
awaiting a merge resolution that will also need to spill their
state.  Having the state directory named 'tg-update' does not
really reflect its purpose.

Instead use the name 'tg-state' for the state spill directory and
add one new state file 'state' that contains the name of the
subcommand that performed the state spill.

Update the places that read/write/check the state to require that
the new state file exist and have the contents "update" in order
for the state to be considered valid.

Essentially this is just a housekeeping update -- there's no effect
on functionality whatsoever.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>