tg.sh: also unset GIT_OBJECT_DIRECTORY when using -C
When using -C and GIT_DIR and/or GIT_COMMON_DIR exists in the
environment, unexpected results occur (it seems as though the
`-C` was ignored).

For that reason `tg.sh` has been unsetting both of those if the
`-C` option gets used.  Now also unset GIT_OBJECT_DIRECTORY too
when `-C` is used as unsetting GIT_DIR but not GIT_OBJECT_DIRECTORY
could produce bizarre and totally unexpected results.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>