README_DOCS.rst: mention source of tg tag -g date and time
The date and time used by `tg tag -g` is NOT the date and time
of the commit referred to by the reflog entry, rather it is
actually the date and time embedded in the reflog entry itself
(which Git does not provide a way to access).

Mention this tidbit of information in the help.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>