README_DOCS.rst: update tg import -s docs
While it's true that the single commit used with the
`tg import -s <name> <commit>` variant can be any commit
and `tg import` itself will not complain, the commit gets
passed to `git cherry-pick` and it will complain about being
given a merge commit.  But it's happy with any non-merge commit
(including a root commit).

Tweak the `tg import` documentation to reflect this.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>