tg-migrate-bases.sh: become POSIX compliant
None of the POSIX specifications for the find utility allow the
`-mindepth` or `-maxdepth` primaries.

However the most recent POSIX specification allows the `-path`
primary (which has generally been supported everywhere that the
`-mindepth` and `-maxdepth` primaries were).

Replace use of `-mindepth` and `-maxdepth` with a substitute that
uses `-path` instead.  It should be almost as efficient.

Since `tg migrate-bases` is not expected to be used very often,
if ever, the minuscule performance hit (an extra fork or maybe two)
won't really matter and in return only POSIX-specified primaries
are used with the find utility.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>