Makefile.sh: sort -u Makefile.dep
Each dependency written to Makefile.dep will be exactly one line.
That makes it amenable to easy sorting.

Although the order of the lines in Makefile.dep does not actually
matter, at some point in the future, a version of Makefile.dep
may be checked in to assist the top-level Makefile in determining
whether or not something is out-of-date.

It will be important at that time to not gratuitously generate
a version of Makefile.dep that differs from the checked-in version.

To that end, start sorting and uniq'ing the lines written to
Makefile.dep now (of course using LC_ALL=C).

Since Git actually checks the hash before deciding whether or not
something has been modified, as long as the exact same list of
dependencies gets generated each time it will not matter that
Makefile.dep has been checked in -- running make in that case will
not produce a dirty worktree.  If it does produce a dirty worktree,
then that suggests a new version of Makefile.dep needs to be checked
in or at the very least the situation investigated.

For the time being, just sort and uniq (the sort `-u` option) the
contents of Makefile.dep to prepare for such a future change.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>