tg-contains.sh: accomodate new bases location
Be careful when parsing the output of fer_branch_contains to make
sure that top-bases stored under refs/heads are not mistakenly
treated as TopGit topic branches rather than their bases.

This can only happen with the new bases location and then only
with just the right combination of circumstances.

Address the problem by discarding any top-bases results immediately
as they are not of interest when determining the TopGit branch
that contains the specified committish.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>