test-lib-tg: grok default --top-bases of tg being tested
Instead of defaulting the value of `$tg_test_bases` to `refs`,
base the default on the result of `tg --top-bases` where the `tg`
being used is the `tg` under test AND the current directory has
been set to the empty directory.

When tg.sh changes its default bases location in the future, the
test library will then continue to work as expected.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>