tg-update: exclude .top* files when using --base
The base branch of an TopGit topic branch should not include any
.topdeps and/or .topmsg files.

The `tg create` command takes extra pains to make sure that the
base of any newly created TopGit topic branch does not have any
.topdeps or .topmsg files at the top level of its tree.

However, the `tg update --base` mode has been using a simple `git
merge` to update the base commit of the specified TopGit topic
branch which will allow .topdeps and/or .topmsg files to infiltrate
the TopGit topic's base branch if the committish the base is being
updated to with `tg update --base` contains them.

Correct this oversight by switching to the same git_topmerge utility
used elsewhere to complete the merge of the new base committish
into the TopGit topic branch and use the same technique that `tg
create` does to check for and expunge any .topdeps and/or .topmsg
files that are present in the committish the base is being updated
to when needed.

Flip the t5001 unit test that was failing from `test_expect_failure`
to `test_expect_success`.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>