t5002: explicitly test the issue 11 problem
The problem described in issue 11 stems from fast-forwarding the
base using `tg update --base` to a commit that contains a .topdeps
and .topmsg file and whose tree matches the tip of the topic branch.

That problem has now been addressed.

Nevertheless, immortalize the essence of the problem report with
its own test that duplicates the issue (and has been confirmed to
fail when run using topgit-0.19.12) but which now passes.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>