tg-export: support --notes option
When exporting in the --collapse or --linearize mode, with --notes
export any comments in the .topmsg file following a '---' separator
to the specified notes ref (refs/notes/commits by default).

Although the default is not to export these comments (they have
always been stripped and discarded in the past), that can be changed
by setting the new config option `topgit.notesExport` to a boolean
or the name of a notes ref.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>