tg-import: pass --no-deps to tg create when appropriate
The `tg create` command supports creating a TopGit topic branch
from an unborn branch as well as a detached HEAD.

If the `tg import` command line specifies a dependency of `HEAD`
(either explicitly, via the `@` alias or implicitly) AND `HEAD` is
either detached or unborn, then automatically supply the `--no-deps`
option to `tg create` when creating the first imported topic branch.

Update t8100 to reflect the new functionality.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>