topgit.html: generate using a consistent style
Eviscerate the guts of the generated topgit.html content and
transplant it into something that produces a more pleasing result.

The resulting final topgit.html documention file should now
always be essentially the same no matter where it's generated.

It is also now color-agnostic and can therefore be viewed in
"dark" mode if desired.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>