README_DOCS.rst: update the out-of-date glossary entry
Correct the bit about any of the .topdeps branches not being
contained by the branch.

It's actually the branch's base that is the merge of all of the
.topdeps branches which means the branch is actually out of
date unless the branch's _base_ contains each and every branch
listed in the branch's .topdeps file.

Add the missing "base" word to the text.

Also tidy up a few bits in the area at the same time.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>