README_DOCS.rst: tweak tg tag -g date comments
Correct the information about Git showing the actual date and time
from the reflog entry itself -- it *is* possible to get Git to show
it, but only at the expense of the index number.

It's of no significance though as that's no help with regards to
showing the reflog of an annotated/signed tag.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>