tg-tag.sh: discard irrelevant spam when using --drop
When running a command such as:

    tg tag --drop HEAD@{3}

The 'HEAD@{3}' part needs to be validated.  When doing so, any
spam that might be emitted to STDERR such as:

    warning: Log for ref HEAD unexpectedly ended on ...

is not relevant to dropping the entry.  Git's own
`git reflog delete` operation does *not* display such spam,
therefore `tg tag --drop` should not either.

Suppress the STDERR output when checking the item about to be
dropped, but if there is a failure, then repeat the operation
to make sure that all error messages show before exiting with
a non-zero status.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>