t6101: perform a few more tg tag --drop tests
When running with a version of Git where <ref> and <ref>@{0}
can be different, test the behavior of dropping <ref>@{0}
when that is the case.

Although Git version 2.31.0 or later is not specifically
required by the test, during setup and verification, if it's
not possible to create a different <ref> and <ref>@{0} value,
the later unit tests will be skipped.

Practically that means Git version 2.31.0 or later is required
for anything to actually be tested by t6101 -- earlier Git
versions will just automatically skip those tests.

Unfortunately, it also means two unit tests (when they are
actually run) are marked `test_expect_failure` due to the
current tg-tag.sh behavior in such a case.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>