tg-tag: do not update ref when dropping @{0} if different
Up until Git v2.31.0, if a reflog existed for <ref> and the
reflog is not empty, the result of `git rev-parse --verify`
for both <ref> and <ref>@{0} are guaranteed to be the same
even if the "@{0}" entry in the reflog does not match <ref>.

However, with Git version 2.31.0 and later, rev-parse will
actually return whatever's in the reflog for "@{0}" which
means it can be different from the <ref> value itself.

If `tg tag --drop` is being asked to drop <ref>@{0} and that
reflog entry does not match the value for <ref>, then do not
pass the `--updateref` argument to `git reflog delete`.

If the values are the same, continue to pass `--updateref` when
dropping "@{0}" just the same as always.

Toggle the two `test_expect_failure` unit tests in t6101 to
`test_expect_success` to reflect this change.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>