tg-update: do not attempt to merge missing depedency
Normally all dependencies must be present in order for the
update to proceed at all.  However, if `--skip-missing` has
been given, the update can proceed with missing dependencies.

When proceeding with missing dependencies, be sure to avoid
trying to merge in dependencies that are missing to avoid
unexpected errors and premature failure for `--skip-missing`.

Update t5011 unit test from `test_expect_failure` to
`test_expect_success` now that it works properly.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>