t5012: test tg update with remote .topdeps removal
When fetching updates from a remote that brings in a .topdeps file
that has had dependency lines removed (and removed properly with
the necessary "revert" patch added to the base) and simultaneously
brings in an update to one of the branches that depends on that
dependency removal, test that the local update succeeds.

Unfortunately one of the unit tests must be marked `test_expect_failure`
at this time.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>