tg-update: avoid updating remote-removed dependencies
When updating a branch that has a remote present, if the remote
branch has removed one or more dependencies, skip them during the
recursive dependencies update stages since they have already been
removed but that will not be reflected in the local .topdeps file
until the final step in the branch update process.

To accomplish this, simply virtually pre-merge the .topdeps file
and compare it to the local version -- if the merge succeeds,
any remote-removed dependencies are then skipped.

If the virtual pre-merge fails, that suggests that the local and
remote versions of the branch disagree about whether or not to
remove the dependency in which case it's not skipped during
recursive dependency updates and the final step in the branch
update process will produce a conflict that must then be manually
resolved.

In the future, it might be possible to resolve the conflict at
the virtual pre-merge stage and then reuse that (if rerere.enabled
is true) resolution automatically at the later stage thereby
skipping any dependencies removed by the manual merge resolution.
The current change is not yet sophisticated enough to do that.

Update the affected t5012 unit test from `test_expect_failure`
to `test_expect_success` now that it works as intended.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>