t5021: test partial fast-forward dependency updates
When updating the local base and the remote branch's base
is merged in as the first dependency and that results in
a fast forward, and there's still at least one other
update to merge in that cannot fast forward; when creating
the merge commit, the first parent must not be the original
commit, it must be the commit that was fast-forwarded to.

Failure to do this correctly results in the remote base
not being contained by the local base although things seem
to work (more or less) because the remote branch contains its
remote base and when that remote branch gets merged into the
local branch that ends up pulling in the remote base changes
into the local branch, but not in proper order.

Add a new test to check for this and, unfortunately, mark
one of the unit tests `test_expect_failure` to reflect that
there's currently an unresolved problem.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>