tg--merging: use correct first parent when fast-forwarding
Correct a long-standing error in the way fast-forwards have been
handled during octopus dependency updates.

During an octopus update, each head being merged in is checked to
see whether or not it represents a fast-forward of the initial
starting point -- if it does, the starting point is simply
fast-forwarded.

If all the heads are contained there's no problem.

Similarly, if none of the heads fast-forward there's no problem.

However, if not all the heads being merged in fast-forward and a
merge commit needs to be created, it's essential that the first
parent of the merge commit be the commit that was fast-forwarded
to and _not_ the original initial starting commit.  Otherwise the
commit that was fast-forwarded to will not be an ancestor like it
should be.  This results in all sorts of unwanted side effects
such as the remote base not being contained by the local base when
the local base fast-forwards to the remote base.

Make the trivial one-line fix to correct this and toggle the
t5021 unit test that checks this from `test_expect_failure` to
`test_expect_success` to reflect the correction.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>