t6102: test tg tag -g with tabs present
Although the exact format of the reflog file lines is detailed
explicitly in `git help update-ref`, it omits any discussion of the
effect of extra tabs on the format and exactly how they are handled.

Since a tab is used as a delimiter between the committer and the
message (if the message is present), it's quite relevant.

Arrange for a whole bunch of tabs to show up in some test commits
and exercise `tg tag -g` with the resulting reflog.

Unfortunately, two of the unit tests must be marked `test_expect_failure`
at this time.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>