README_DOCS.rst: wordsmith some of the commentary
It's not relevent _why_ Git fails (and, in this case, the analysis
either was wrong or later became wrong); what's relevant is that
Git has a major FAIL when dealing with reflogs of annotated/signed
tags.

It appears that the current excuse is that `git reflog show` being
just an alias for `git log -g --abbrev-commit --pretty=oneline` --
according to `git help reflog` -- means that anything other than a
commit cannot be handled since `git log` only deals with commits.

In any case the *WHY* is not important here, just the *FAIL* bit.
Take out the analysis (which might become incorrect in the future)
and just leave the relevant stuff.  Do add a mention of
`git reflog show` though as not everyone may be aware that it's
just an alias for `git log -g`.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>