tg--merging: manually update MERGE_HEAD
Since Git v2.46.0 and the commit 993d57eded ("refs: pseudorefs are no
refs"), it is no longer allowed to use 'git update-ref MERGE_HEAD'.

Because of this, TopGit was printing this message in case of merge
conflicts:

  Auto-merging foo/bar
  CONFLICT (content): Merge conflict in foo/bar
  fatal: update_ref failed for ref 'MERGE_HEAD': refusing to update pseudoref 'MERGE_HEAD'
  Automatic merge failed; fix conflicts and then commit the result.

But resolving the conflicts had no effects: a 'tg update --continue'
command would go back to the same conflict over and over, because a
simple commit was created instead of a merge one, linked the topic with
its base after the merge resolution.

A simple workaround is to directly write the reference in the
'.git/MERGE_HEAD' file. This way, it should work with previous and
recent Git versions.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>