tg.sh: provide utility function update_dotgit_ref
Beginning with Git v2.46.0, `git update-ref` simply refuses to
update certain ref names such as "MERGE_HEAD" and "FETCH_HEAD" even
though it continues to be able to update others such as "CHERRY_PICK_HEAD".

Apparently as of Git v2.46.0, there's no alternative git command
that can be used to update "MERGE_HEAD" or "FETCH_HEAD".  It's
simply no longer possible via the `git` executable (except as an
indirect result of some Git operation).

Emulate the `git update-ref <refname> <oid>` functionality with a
new `update_dotgit_ref` function that directly writes the resolved
<oid> (verifying that it also exists in the repository) to the
"--git-dir" directory as the file named exactly <refname>.

Of course it's written with a single trailing newline to be in the
proper format.

Do restrict possible <refname> values to only those that consist
only of UPPER ALPHAnumericunderscore characters where the first
must be an ALPHA and the last must not be an underscore.  Also
restrict the length to at least 4 and at most 22 characters for
now.  For the moment, this is sufficient.

This would seem to be moving in the wrong direction by forcing any
scripts/utilities that need to update these ref values to write
them directly to the correctly named file rather than using a `git`
command to do so and thus leaving such scripts/utilities in a more
fragile state and prone to breakage than they were before the poorly
thought out change to the `git update-ref` semantics introduced in
Git version 2.46.0.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>