t/Makefile.mak: `wait` after `set -m` for parallel tests
The combination of using `set -m` and running tests in parallel and
using GNU Make (non-GNU make [e.g. `bmake`/`pmake`] is not susceptible)
can result in the testing process being suspended by a SIGTTIN
signal (to readers pgrp upon background tty read)!

Exactly why this happens, but only with GNU make is not totally clear.

However, the simple expedience of adding a `wait` in the same shell
wherever `set -m` has been used after all the parallel jobs have been
started and before exiting that shell seems to remedy the issue.

Therefore, add a `wait` to the suitable spot in the three places
where `set -m` gets used.

Regardless of the test results, the `wait` is always performed and
then the previous status is restored.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>