t/aggregate-results.sh: exit with non-zero status if failed tests
If there are any failed tests, exit with a non-zero status.

If there are no failed tests, exit with a zero status (as has always
previously been the case).

By having aggregate-results.sh exit with a non-zero status whenever
there are failed tests, the make rule that runs it will then fail.

The reason this is done in aggregate-results.sh rather than failing
the test process earlier when there are failed tests is to make
sure that the aggregate results are actually computed and shown
before stopping because of the failures.

The aggregate-results.sh script is only used by the t/Makefile.mak
"test" target (which does NOT require `prove`) -- it is not used
by the t/Makefile.mak "prove" target at all.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>