Makefile: avoid accidents with environment variable "T"
The "T" makefile variable can be set to restrict the set of tests
to be run to only those tests matching the "glob" that "T" has
been set to (it may be set to multiple space-separated globs).

This restricted set is only intended to be available via the:

    make T="<some_glob>..." test

syntax.  It is NOT to be available like so:

    T="<some_glob>..." make test

nor like so:

    T="<some_glob..."; export T; make test

To prevent an environment variable version of "T" from being picked
up by the top-level make, a "T =" line was added in commit
88ea1a9f0d37356d "Makefiles: minor clean ups and avoid environment
contamination" (2022-12-21).

While, indeed, that works with GNU make (because if T is already
exported into the environment when GNU make runs WITHOUT the -e
option, then when the Makefile sets T to "", it remains exported),
but it does NOT work with some other makes (the "T =" setting is
available inside the makefile as "$(T)", but is NOT exported to
subshells which continue to inherit any "T" from the environment
the original "make" was run in).

Compensate for this the same way the "MAKE" variable was handled
by explicitly exporting the value of "$(T)" as "T" into the environment
whenever running a sub make.

While there are other variables that could also be inherited by
non-GNU sub makes, they are not as likely to be as troublesome as
the "T" environment variable and hence do not (yet) get the same
special treatment.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>