t/Makefile.sh: make sure T is expanded early
If T has been set to specific tests via one or more "glob"s, it's
important to expand those globs before $(T) is used on the left
hand side of a target rule since whether or not filename/pathname
expansion occurs in target rules in makefiles is unspecified.

If T is not a glob (or multiple globs), but an actual filename (or
multiple filenames), it will be retained as-is.

Force T to be set as a $(MAKE) argument when invoking a subshell
to preserve the expansion from Makefile.sh.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>