README_DOCS.rst: add tg update --continue/--abort/--skip/--stop docs
Add a section to the `tg update` command documentation that details
the four options that can be used to proceed after a `tg update`
command stops because it encounters a merge conflict.

While there's already minimal help for these shown when the merge
conflict occurs and in the `tg status` output when a `tg update`
has been interrupted due to a merge conflict, those are extremely
short descriptions.

Provide much more detailed detailed descriptions in the
`tg help update` output.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>