README_DOCS.rst: update rest of tg prev and tg next docs
When the previous to next overhaul was implemented (9d1e18b752656d75
"tg: previous to next overhaul", 2017-03-31, topgit-0.19.7), the
documentation for `tg next`, `tg prev` and `tg checkout` was NOT
updated to reflect the new semantics.

An attempt to correct this was made in 693e14345cbe0cf8 "README_DOCS.rst:
update tg prev and tg next usage summary" (2021-08-23, topgit-0.19.13),
but `tg checkout` was overlooked (it has a "next" and "prev" subcommands).

Update the `tg checkout` documentation for the "next" and "prev"
subcommands at last to finally describe the new "next" and "prev"
semantics that were introduced in topgit-0.19.7.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>